Refactor: 검색 결과 리스트 컴포넌트로 분리
